+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Fixed fatal error after create screenshot or report on Windows
 * Implemented mechanism of applying 3rd party credentials
+* Fixed setValue() method for StrictselectElement class
 
 1.0.0-rc.17
 =============